+++ src/com/scottyplunkett/server/Cycle/Request/HTTPRequest.java
@@ -1,4 +1,4 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
