+++ src/com/scottyplunkett/server/Cycle/Response/Behavior/Handlers/DirectoryContent.java
@@ -1,4 +1,4 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Response.Behavior.Handlers;
 
 import java.io.File;
 import java.nio.file.Path;
