+++ src/com/scottyplunkett/server/Cycle/Response/Behavior/Handlers/HTTPResponseBody.java
@@ -1,4 +1,8 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Response.Behavior.Handlers;
+
+import com.scottyplunkett.server.Cycle.DirectoryContent;
+import com.scottyplunkett.server.Cycle.Response.DirectoryContent;
+import com.scottyplunkett.server.Cycle.Router;
 
 import java.io.IOException;
 import java.nio.file.Files;
