+++ src/com/scottyplunkett/server/Cycle/Response/Behavior/Handlers/ImageHandler.java
@@ -1,4 +1,9 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Response.Behavior.Handlers;
+
+import com.scottyplunkett.server.Cycle.Date;
+import com.scottyplunkett.server.Cycle.Router;
+import com.scottyplunkett.server.Date;
+import com.scottyplunkett.server.HTTPResponseHeaders;
 
 import java.io.IOException;
 import java.nio.file.Files;
