+++ src/com/scottyplunkett/server/Cycle/Response/HTTPResponse.java
@@ -1,4 +1,7 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Response;
+
+import com.scottyplunkett.server.Cycle.Request.HTTPRequest;
+import com.scottyplunkett.server.Cycle.Response.Behavior.Handlers.FormHandler;
 
 import java.io.IOException;
 import java.nio.file.Files;
