+++ src/com/scottyplunkett/server/Cycle/Response/Routing/Router.java
@@ -1,4 +1,4 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Response.Routing;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
