+++ src/com/scottyplunkett/server/Cycle/Utils/Date.java
@@ -1,4 +1,4 @@
-package com.scottyplunkett.server;
+package com.scottyplunkett.server.Cycle.Utils;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
