+++ test/com/scottyplunkett/server/CookieHandlerTest.java
@@ -1,12 +1,15 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.CookieHandler;
+import com.scottyplunkett.server.Cycle.HTTPRequest;
+import com.scottyplunkett.server.Cycle.HTTPResponseHeaders;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
-import static com.scottyplunkett.server.ByteArraysReducer.merge;
+import static com.scottyplunkett.server.Cycle.ByteArraysReducer.merge;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
 class CookieHandlerTest {
