+++ test/com/scottyplunkett/server/DateTest.java
@@ -1,5 +1,6 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.Date;
 import org.junit.jupiter.api.Test;
 
 import java.text.SimpleDateFormat;
