+++ test/com/scottyplunkett/server/DirectoryContentTest.java
@@ -1,5 +1,6 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.DirectoryContent;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Paths;
