+++ test/com/scottyplunkett/server/HTTPResponseBodyTest.java
@@ -1,5 +1,6 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.HTTPResponseBody;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
