+++ test/com/scottyplunkett/server/LogsHandlerTest.java
@@ -1,5 +1,8 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.HTTPRequest;
+import com.scottyplunkett.server.Cycle.HTTPResponseHeaders;
+import com.scottyplunkett.server.Cycle.LogsHandler;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
@@ -8,7 +11,7 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
-import static com.scottyplunkett.server.ByteArraysReducer.merge;
+import static com.scottyplunkett.server.Cycle.ByteArraysReducer.merge;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
 class LogsHandlerTest {
