+++ test/com/scottyplunkett/server/PageContentTest.java
@@ -1,5 +1,7 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.PageContent;
+import com.scottyplunkett.server.Cycle.Router;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
