+++ test/com/scottyplunkett/server/RouterTest.java
@@ -1,5 +1,6 @@
 package com.scottyplunkett.server;
 
+import com.scottyplunkett.server.Cycle.Router;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
